Handle table/lookupswitch
Not failing on printing should have been enough,
but in the exceptionally rare case someone conjured bytecode where something was pushed to stack and a switch is before the NPE, try handle it.

It works because all branch targets must enter and exit with the same stack shape, so each switch target may push to stack, but must pop equally.
Just brute-force walking through the bytecode will arrive at the switch opcode which itself pops 1 slot from stack. Past this point it's all the same.

If the target is pushed in one of the switch targets, this analysis will point out the last branch in bytecode which similarly to ternary operator result method chaining may produce the wrong npe source, but can live with it.